+++ glossary/glossary-service-phrase/src/main/java/com/systema/glossary/service/phrase/PhraseService.java
@@ -1,6 +1,7 @@
 package com.systema.glossary.service.phrase;
 
 import java.util.List;
+import java.math.BigDecimal;
 
 import com.systema.glossary.model.phrase.Phrase;
 
@@ -8,11 +9,11 @@ public interface PhraseService {
 	
 	public Phrase createPhrase(Phrase phrase);
 	
-	public void deletePhrase(Long id);
+	public void deletePhrase(BigDecimal id);
 	
 	public List<Phrase> getAllPhrases();
 	
-	public Phrase getPhraseById(Long id);
+	public Phrase getPhraseById(BigDecimal id);
 	
 	public List<Phrase> getPhraseListByTextMatch(String searchText);
